Python Jupyter notebook
Our project is about digging into a big set of data from English Premier League football matches to find patterns and make predictions about which teams will win. Football is not just a game for us; it's a puzzle with pieces scattered in rows of data. With over 5000 matches' worth of details, we're looking to sort through the stats and come up with smart guesses on game outcomes. It's a project that mixes our love of the sport with our interest in data, and we hope it can help fans and experts get new insights into the game.